+++ RELEASES.md
@@ -76,6 +76,7 @@ Upcoming
 - Upgrade commons-fileupload (fix file name always in lower case)
 - Get configuration name from file name when configuration is uploaded
 - Store configuration version and file name too when configuration is uploaded
+- Add TimeOutException to IbisLocalSender (next to regular SenderException)